chore: moves server files from fileglancer_server/ to fileglancer/
also updates fileglancer_server to fileglancer inside the files, and changes relative imports to absolute (e.g., fileglancer.filestore instead of .filestore
