+++ fileglancer/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Python unit tests for fileglancer."""
